pref_data_generation/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=1.1.1=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiosignal=1.3.2=pypi_0
airportsdata=20250224=pypi_0
albucore=0.0.13=pypi_0
albumentations=1.4.10=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.6.2.post1=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
astor=0.8.1=pypi_0
asttokens=3.0.0=pypi_0
async-lru=2.0.4=pypi_0
attrs=24.3.0=pypi_0
av=13.1.0=pypi_0
babel=2.16.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blake3=1.0.4=pypi_0
bleach=6.2.0=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.9.24=h06a4308_0
cachetools=5.5.2=pypi_0
certifi=2024.8.30=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.1.1=pypi_0
cobble=0.1.4=pypi_0
comm=0.2.2=pypi_0
compressed-tensors=0.9.2=pypi_0
contourpy=1.3.1=pypi_0
cryptography=43.0.3=pypi_0
cupy-cuda12x=13.4.1=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.11=pypi_0
datasets=3.2.0=pypi_0
debugpy=1.8.11=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
depyf=0.18.0=pypi_0
diffusers=0.31.0=pypi_0
dill=0.3.8=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
dnspython=2.7.0=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.8.0=pypi_0
email-validator=2.2.0=pypi_0
executing=2.1.0=pypi_0
face-detection=0.2.2=pypi_0
fastapi=0.115.12=pypi_0
fastapi-cli=0.0.7=pypi_0
fastjsonschema=2.21.1=pypi_0
fastrlock=0.8.3=pypi_0
filelock=3.16.1=pypi_0
fire=0.7.0=pypi_0
flash-attn=2.7.0.post2=pypi_0
flask=3.1.0=pypi_0
fonttools=4.55.0=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.9.0=pypi_0
gguf=0.10.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
google-search-results=2.4.2=pypi_0
greenlet=3.1.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httptools=0.6.4=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.26.2=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.10=pypi_0
imageio=2.36.0=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=8.5.0=pypi_0
interegular=0.3.3=pypi_0
iopath=0.1.10=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.30.0=pypi_0
ipywidgets=8.1.5=pypi_0
isoduration=20.11.0=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
jiter=0.8.0=pypi_0
joblib=1.4.2=pypi_0
json5=0.10.0=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter=1.1.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.11.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.2=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.3.3=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kiwisolver=1.4.7=pypi_0
langchain=0.3.17=pypi_0
langchain-core=0.3.33=pypi_0
langchain-text-splitters=0.3.5=pypi_0
langsmith=0.3.3=pypi_0
lark=1.2.2=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
llava=1.7.0.dev0=pypi_0
llguidance=0.7.11=pypi_0
llvmlite=0.43.0=pypi_0
lm-format-enforcer=0.10.11=pypi_0
lmdb=1.5.1=pypi_0
loguru=0.7.2=pypi_0
lxml=5.3.0=pypi_0
mammoth=1.8.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markdownify=0.14.1=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.2=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mistral-common=1.5.4=pypi_0
mistune=3.0.2=pypi_0
modelscope=1.24.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
msgspec=0.19.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
nbclient=0.10.1=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.4.2=pypi_0
ninja=1.11.1.4=pypi_0
nltk=3.9.1=pypi_0
notebook=7.3.1=pypi_0
notebook-shim=0.2.4=pypi_0
numba=0.60.0=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-cusparselt-cu12=0.6.2=pypi_0
nvidia-ml-py=12.560.30=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
omegaconf=2.3.0=pypi_0
openai=1.55.1=pypi_0
opencv-contrib-python=4.10.0.84=pypi_0
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openssl=1.1.1w=h7f8727e_0
opt-einsum=3.3.0=pypi_0
orjson=3.10.15=pypi_0
outlines=0.1.11=pypi_0
outlines-core=0.1.26=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
paddleocr=2.9.1=pypi_0
paddlepaddle-gpu=2.6.2=pypi_0
pandas=2.2.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
partial-json-parser=0.2.1.1.post4=pypi_0
pathvalidate=3.2.1=pypi_0
pdfminer=20191125=pypi_0
pdfminer-six=20240706=pypi_0
peft=0.13.2=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py311h06a4308_0
platformdirs=4.3.6=pypi_0
portalocker=3.0.0=pypi_0
prometheus-client=0.21.0=pypi_0
prometheus-fastapi-instrumentator=7.1.0=pypi_0
prompt-toolkit=3.0.48=pypi_0
propcache=0.2.0=pypi_0
protobuf=5.28.3=pypi_0
psutil=6.1.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
puremagic=1.28=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyairports=2.1.1=pypi_0
pyarrow=18.1.0=pypi_0
pyav=13.1.0=pypi_0
pyclipper=1.3.0.post6=pypi_0
pycountry=24.6.1=pypi_0
pycparser=2.22=pypi_0
pycryptodome=3.21.0=pypi_0
pydantic=2.10.2=pypi_0
pydantic-core=2.27.1=pypi_0
pydub=0.25.1=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.2.0=pypi_0
python=3.11.0=h7a1cb2a_3
python-dateutil=2.9.0.post0=pypi_0
python-docx=1.1.2=pypi_0
python-dotenv=1.0.1=pypi_0
python-json-logger=3.2.1=pypi_0
python-multipart=0.0.20=pypi_0
python-pptx=1.0.2=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=pypi_0
qwen-vl-utils=0.0.8=pypi_0
rapidfuzz=3.10.1=pypi_0
ray=2.44.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-toolbelt=1.0.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.9.4=pypi_0
rich-toolkit=0.14.0=pypi_0
rpds-py=0.21.0=pypi_0
safetensors=0.4.5=pypi_0
sam-2=1.0=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
send2trash=1.8.3=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.24.1=pypi_0
serpapi=0.1.5=pypi_0
setproctitle=1.3.5=pypi_0
setuptools=75.1.0=py311h06a4308_0
shapely=2.0.6=pypi_0
shellingham=1.5.4=pypi_0
shortuuid=1.0.13=pypi_0
six=1.16.0=pypi_0
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlalchemy=2.0.37=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
starlette=0.46.1=pypi_0
sympy=1.13.1=pypi_0
tenacity=9.0.0=pypi_0
termcolor=2.5.0=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.9.20=pypi_0
tiktoken=0.9.0=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.1=pypi_0
tomli=2.1.0=pypi_0
torch=2.6.0=pypi_0
torchaudio=2.6.0=pypi_0
torchvision=0.21.0=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.50.2=pypi_0
triton=3.2.0=pypi_0
typer=0.15.2=pypi_0
types-python-dateutil=2.9.0.20241206=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.3=pypi_0
uvicorn=0.34.0=pypi_0
uvloop=0.21.0=pypi_0
vllm=0.8.2=pypi_0
wandb=0.19.8=pypi_0
watchfiles=1.0.4=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.11.1=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
websockets=14.1=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.44.0=py311h06a4308_0
widgetsnbextension=4.0.13=pypi_0
xformers=0.0.29.post2=pypi_0
xgrammar=0.1.16=pypi_0
xlsxwriter=3.2.0=pypi_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.18.3=pypi_0
youtube-transcript-api=0.6.3=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
zstandard=0.23.0=pypi_0
